+++ public/manifest.json
@@ -1,5 +1,5 @@
 {
-  "short_name": "React App",
+  "short_name": "Dev Portfolio",
   "name": "Create React App Sample",
   "icons": [
     {
