--- src/components/header/Dot.css
@@ -1,32 +1,0 @@
-.dot {
-  display: block;
-  width: var(--dot-size);
-  height: var(--dot-size);
-  background: gray;
-  border-radius: 50%;
-  opacity: 0.3;
-
-  -webkit-transition: all 0.5s;
-  -ms-transition: all 0.5s;
-  transition: all 0.5s;
-}
-
-@keyframes animateDotIn {
-  0% {
-    opacity: 0;
-    transform: scale(0.6) translateY(-8px) translateX(-8px);
-  }
-
-  100% {
-    opacity: 0.3;
-  }
-}
-
-@keyframes fader {
-  0% {
-    opacity: 30%;
-  }
-  100% {
-    opacity: 5%;
-  }
-}
